+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/navigation/ComFinNavGraph.kt
@@ -27,14 +27,14 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import br.com.rbrthmn.home.ui.HomeDestination
 import br.com.rbrthmn.home.ui.HomeScreen
-import br.com.rbrthmn.misc.incomedivisions.IncomeDivisionsDestination
-import br.com.rbrthmn.misc.incomedivisions.IncomeDivisionsScreen
-import br.com.rbrthmn.misc.morefeatures.MoreFeaturesDestination
-import br.com.rbrthmn.misc.morefeatures.MoreFeaturesScreen
-import br.com.rbrthmn.misc.recurringexpenses.RecurringExpenses
-import br.com.rbrthmn.misc.recurringexpenses.RecurringExpensesDestination
-import br.com.rbrthmn.misc.reserves.ReservesDestination
-import br.com.rbrthmn.misc.reserves.ReservesScreen
+import br.com.rbrthmn.misc.ui.incomedivisions.IncomeDivisionsDestination
+import br.com.rbrthmn.misc.ui.incomedivisions.IncomeDivisionsScreen
+import br.com.rbrthmn.misc.ui.morefeatures.MoreFeaturesDestination
+import br.com.rbrthmn.misc.ui.morefeatures.MoreFeaturesScreen
+import br.com.rbrthmn.misc.ui.recurringexpenses.RecurringExpenses
+import br.com.rbrthmn.misc.ui.recurringexpenses.RecurringExpensesDestination
+import br.com.rbrthmn.misc.ui.reserves.ReservesDestination
+import br.com.rbrthmn.misc.ui.reserves.ReservesScreen
 import br.com.rbrthmn.operations.ui.OperationsDestination
 import br.com.rbrthmn.operations.ui.OperationsScreen
 import br.com.rbrthmn.settings.SettingsDestination
